backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
rdkit-pypi>=2023.9.1
